+++ packages/Contracts/src/Database/Query/Exceptions/CriteriaException.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Aedart\Contracts\Database\Query\Exceptions;
+
+use Throwable;
+
+/**
+ * Criteria Exception
+ *
+ * General criteria exception.
+ *
+ * @author Alin Eugen Deac <ade@rspsystems.com>
+ * @package Aedart\Contracts\Database\Query\Exceptions
+ */
+interface CriteriaException extends Throwable
+{
+}